+++ config/routes.rb
@@ -66,6 +66,5 @@ Mosh::Application.routes.draw do
   
   resources :home
 
-
   root to: 'home#index'
 end
