+++ db/migrate/20130611022616_create_images.rb
@@ -1,13 +1,10 @@
-class CreateImages < ActiveRecord::Migration
-  def change
-    create_table :images do |t|
-      t.string :name
-      t.datetime :solicit_start
-      t.datetime :submit_deadline
-      t.datetime :test_deadline
-      t.integer :install_id
-
-      t.timestamps
-    end
+class CreateFiles < ActiveRecord::Migration
+def change
+  create_table :files do |t|
+  t.string :name
+  # If using MySQL, blobs default to 64k, so we have to give
+  # an explicit size to extend them
+  t.binary :data, :limit => 1.megabyte
   end
+end
 end
