+++ test/unit/helpers/home_helper_test.rb
@@ -1,0 +1,4 @@
+require 'test_helper'
+
+class HomeHelperTest < ActionView::TestCase
+end
