--- test/unit/helpers/test_home_helper_test.rb
@@ -1,4 +1,0 @@
-require 'test_helper'
-
-class TestHomeHelperTest < ActionView::TestCase
-end
